ATOM-5921 Material Editor: Select newly created materials in the asset browser
The code was previously using asset browser notifications to listen for new files being added in order to select newly created materials.
Attempting to change the selection within the notification failed because the new entry still had not been added.
Now the material browser queues and processes the selection on tick.

https://jira.agscollab.com/browse/ATOM-5921