ATOM-14040 Add Support for Cavity Maps
Before working on adding cavity map support, I am replacing the 010_SpecularOcclusion test case with maps that make it easier to distinguish what's going on. The brick map was just too noisy. This simpler tile map will make it easier to see if there are artifacts like banding that get introduced somewhere along the way.

There are corresponding changes in AtomSampleViewer for the baseline screenshot.