Replaced GetAllScenes in AzFramework::Scenes with two new functions that use a callback.
The callback is a more performant version compared to the original version that needed to do a copy. The second function allows keeping track of scenes that have already been deleted. Not all scenes can be immediately deleted and systems may hold on to it until they can. This can for instance happen if a system is waiting for an async call to complete or needs to wait till hardware resources are made available. These systems may still require periodic ticking or other interactions. To support this zombie scenes are tracked until they have no more references.