[1.0 only] Fix context menu behavior in the old Editor viewport
When calling exec with the old input implementation, input events continue getting sent to the viewport and unexpected behavior can occur. For 1.0 only, we're switching this to use popup again (which has its own issues when run against the new viewport).

NOTE: This is for 1.0 only, please do not cherry-pick this to main!