ATOM-15372 Restoring material browser filter logic
Restoring material browser filter logic to only show folders with content that matches the filter.

This change was originally introduced with the functionality to create a new folder within the material browser.  Creating a new, empty folder would not display the folder in the browser because it did not have any content that matched the filter.  While that makes sense to address the issue with the create folder command, it makes it extremely difficult and navigate the browser to find assets that match the filter because every folder that does not match the filter is also displayed in the tree.  We can revisit issues with the create folder action later.

https://jira.agscollab.com/browse/ATOM-15372