Added a generation step within the Projects.cmake file to generate a .setreg file containing the CMake build directory root. The file is output to the <project-root>/user/Registry/build_path.setreg file. This occurs only on non-host platforms when configuring for non-Monolithic builds, since the Build Directory is used to located the project binary directory in order to located the project's generated ${CMAKE_BINARY_DIR}/bin/$<CONFIG>/Registry directory containing the cmake_dependencies.<project-name>.<application-name>.setreg file containing the list of gem modules to load for a given application
Updated the SettingsRegistryMergeUtils AddRuntimeFilePaths function to
be read in the new "/Amazon/Project/Settings/Build/project_build_path"
and use that to form an absolute path to the project build directory by
appending it to the FilePathKey_ProjectPath key.
That key is set in the 'FilePathKey_ProjectBuildPath' constant
Next updated the SettingsRegistryMergeUtils
MergeSettingsToRegistry_TargetBuildDependencyRegistry function to look
within the project build directory to locate the
cmake_dependencies.*.setreg file to load

Tweaked the Settings Registry merge order of the ComponentApplication,
GameApplication and Settings Registry builder to merge the command line
after merging the global user registry and after merging the project
user registry.
Moved the call to MergeSettingsToRegistry_TargetBuildDependencyRegistry
to occur after the above calls to make sure the properly overriden
projects' user registry was merged in order for the correct project
build path to be stored in the SettingsRegistry