Preserve asset ids for assets that fail to load, when deserializing from json
Sometimes deserializing a Json document happens when asset handlers are not
registered. In that case, `FindOrCreateAsset` will fail to create the
asset, since there's no handler registered to create it. When this happens,
`FindOrCreateAsset` returns an Asset instance with a null asset id. This
effectively causes the json deserializer to lose that data, even in
situations where the the actual asset data doesn't need to be loaded, but
the asset id needs to be preserved.