Fix locating project module for UnixLike platforms
Fixes the issue with project-centric workflows running GameLauncher, and
it opens AP which can't find the project dynamic module.  From AP's
perspective, the project module is not in the executable directory,
which is in engine bin.  The SystemFile::Exists check is true on the
file because it uses the 'cwd'.  In that situation, an absolute path
must be obtained for the module to be loaded.