Project dll is not loaded by the AP when opened from the launcher (#1123)
* Fixes locating the project dll when using SDK

SDK engine usage has project dll in the project build path, but
searching for module filepaths for loading would have a passing
SystemFile::Exists check but no full filepath was amended to the module.
This causes the module to fail to load.

* Fix locating project module for UnixLike platforms

Fixes the issue with project-centric workflows running GameLauncher, and
it opens AP which can't find the project dynamic module.  From AP's
perspective, the project module is not in the executable directory,
which is in engine bin.  The SystemFile::Exists check is true on the
file because it uses the 'cwd'.  In that situation, an absolute path
must be obtained for the module to be loaded.

* Add missing header to fix UnixLike builds

* Applies a suggested change from PR

Use operator-> on the AZStd::optional

* Add semicolon to a class macro line

Prevent auto formatting indenting the following line.