[SPEC-5991] Removing ImageHDR.h/.cpp. These files are no longer needed and created a licensing issue. (#1420)
The only place this code was referenced was in using textured tooltips, if the texture used for that tooltip was a .hdr file. However, the few .hdr files in all of o3de are Atom material / lighting related - definitely not used for tooltips.

It's likely bitmap tooltips in general are not used at all anymore, so more code could probably be ripped out, but this solves the immediate problem without making too many changes.