If there are too many wrinkle masks, discard the lowest weight ones (#988)
Since the shader supports a fixed maximum number of active wrinkle masks, if the number of active wrinkle masks is larger than that, it will crash where wrinkleMaskObjectSrg->SetImageArray is called with an array_view which is larger than the target image array.

We need a sensible method of truncating the number of masks. One such method is to discard the lowest weight masks.